+++ dev/ci/user-overlays/14059-ejgallego-coq+path_boot.sh
@@ -1,0 +1,1 @@
+overlay elpi https://github.com/ejgallego/coq-elpi coq+path_boot 14059