+++ ide/coqide/unicode_bindings.ml
@@ -48,9 +48,11 @@ let load_files filenames =
   let add f =
     selected_filenames := f::!selected_filenames in
   let warn_default_not_found () =
-    Ideutils.warning (Printf.sprintf
-      "Warning: the file 'ide/default.bindings' was not found in %s."
-      (Envars.coqlib())) in
+    let dirs = Minilib.coqide_data_dirs () in
+    Ideutils.warning Format.(
+        asprintf
+          "@[Warning: the file 'ide/default.bindings' was not found in:@\n @[<v>%a@]@]."
+          (pp_print_list ~pp_sep:pp_print_cut pp_print_string) dirs) in
   let warn_local_not_found () =
     Ideutils.warning (Printf.sprintf
       "Warning: the local configuration file 'coqide.bindings' was not found.") in