+++ lib/envars.ml
@@ -115,44 +115,6 @@ let _ =
 let use_suffix prefix suffix =
   if String.length suffix > 0 && suffix.[0] = '/' then suffix else prefix / suffix
 
-(** [check_file_else ~dir ~file oth] checks if [file] exists in
-    the installation directory [dir] given relatively to [coqroot],
-    which maybe has been relocated.
-    If the check fails, then [oth ()] is evaluated.
-    Using file system equality seems well enough for this heuristic *)
-let check_file_else ~dir ~file oth =
-  let path = use_suffix coqroot dir in
-  if Sys.file_exists (path / file) then path else oth ()
-
-let guess_coqlib fail =
-  getenv_else "COQLIB" (fun () ->
-  let prelude = "theories/Init/Prelude.vo" in
-  check_file_else ~dir:Coq_config.coqlibsuffix ~file:prelude
-    (fun () ->
-      if Sys.file_exists (Coq_config.coqlib / prelude)
-      then Coq_config.coqlib
-      else
-        fail "cannot guess a path for Coq libraries; please use -coqlib option \
-              or ensure you have installed the package containing Coq's stdlib (coq-stdlib in OPAM) \
-              If you intend to use Coq without a standard library, the -boot -noinit options must be used.")
-  )
-
-let coqlib_ref : string option ref = ref None
-let set_user_coqlib path = coqlib_ref := Some path
-
-(** coqlib is now computed once during coqtop initialization *)
-
-let set_coqlib ~fail =
-  match !coqlib_ref with
-  | Some _ -> ()
-  | None ->
-    let lib = guess_coqlib fail in
-    coqlib_ref := Some lib
-
-let coqlib () = Option.default "" !coqlib_ref
-let coqcorelib () =
-  getenv_else "COQCORELIB" (fun () -> Option.cata (fun c -> Filename.concat c "../coq-core/") "" !coqlib_ref)
-
 let docdir () =
   (* This assumes implicitly that the suffix is non-trivial *)
   let path = use_suffix coqroot Coq_config.docdirsuffix in
@@ -206,9 +168,12 @@ let xdg_dirs ~warn =
 (* Print the configuration information *)
 
 let print_config ?(prefix_var_name="") f =
+  let env = Boot.Env.init () in
+  let coqlib = Boot.Env.coqlib env |> Boot.Path.to_string in
+  let corelib = Boot.Env.corelib env |> Boot.Path.to_string in
   let open Printf in
-  fprintf f "%sCOQLIB=%s/\n" prefix_var_name (coqlib ());
-  fprintf f "%sCOQCORELIB=%s/\n" prefix_var_name (coqlib () / "../coq-core/");
+  fprintf f "%sCOQLIB=%s/\n" prefix_var_name coqlib;
+  fprintf f "%sCOQCORELIB=%s/\n" prefix_var_name corelib;
   fprintf f "%sDOCDIR=%s/\n" prefix_var_name (docdir ());
   fprintf f "%sOCAMLFIND=%s\n" prefix_var_name (ocamlfind ());
   fprintf f "%sCAMLFLAGS=%s\n" prefix_var_name Coq_config.caml_flags;