+++ topbin/coqnative_bin.ml
@@ -229,15 +229,14 @@ let add_rec_path ~unix_path ~coq_root =
     Feedback.msg_warning (str "Cannot open " ++ str unix_path)
 
 let init_load_path_std () =
-  let () = Envars.set_coqlib ~fail:(fun x -> CErrors.user_err Pp.(str x)) in
-  let ( / ) = Filename.concat in
-  let coqlib = Envars.coqlib () in
-  let user_contrib = coqlib/"user-contrib" in
+  let env = Boot.Env.init () in
+  let stdlib = Boot.Env.stdlib env |> Boot.Path.to_string in
+  let user_contrib = Boot.Env.user_contrib env |> Boot.Path.to_string in
   let xdg_dirs = Envars.xdg_dirs in
   let coqpath = Envars.coqpath in
   (* NOTE: These directories are searched from last to first *)
   (* first standard library *)
-  add_rec_path ~unix_path:(coqlib/"theories") ~coq_root:(Names.DirPath.make[coq_root]);
+  add_rec_path ~unix_path:stdlib ~coq_root:(Names.DirPath.make[coq_root]);
   (* then user-contrib *)
   if Sys.file_exists user_contrib then
     add_rec_path ~unix_path:user_contrib ~coq_root:Loadpath.default_root_prefix;
@@ -350,7 +349,7 @@ let rec parse_args (args : string list) accu =
   | "-coqlib" :: s :: rem ->
     if not (Minisys.exists_dir s) then
       fatal_error (str "Directory '" ++ str s ++ str "' does not exist") false;
-    Envars.set_user_coqlib s;
+    Boot.Env.set_coqlib s;
     parse_args rem accu
   | ("-?"|"-h"|"-H"|"-help"|"--help") :: _ -> Usage.usage ()
   | [file] ->